finished integration between datastore and collector. TODO: history records overwrite each other, needs to be fixed

+++ PsDataStoreDb/dashboard2/src/main/java/gov/bnl/racf/ps/dashboard/db/object_manipulators/PsObjectShredder.java
@@ -12,25 +12,28 @@ import org.hibernate.Session;
 
 /**
  * utility class for deleting objects
+ *
  * @author tomw
  */
 public class PsObjectShredder {
+
     /**
      * delete host object
+     *
      * @param session
-     * @param host 
+     * @param host
      */
-    public static void delete(Session session, PsHost host){
+    public static void delete(Session session, PsHost host) {
         //first order of business is to remove this host from any
         // sites it might belong to
         List listOfAllSites = PsDataStore.getAllSites(session);
         Iterator iter = listOfAllSites.iterator();
-        while(iter.hasNext()){
-            PsSite currentSite = (PsSite)iter.next();
+        while (iter.hasNext()) {
+            PsSite currentSite = (PsSite) iter.next();
             currentSite.removeHost(host);
         }
-        
-        
+
+
         // finally delete the host itself
         session.delete(host);
         //TODO add deleting services associated with this host
@@ -40,6 +43,7 @@ public class PsObjectShredder {
         //throw new UnsupportedOperationException("Not yet implemented");
         session.delete(service);
     }
+
     public static void delete(Session session, PsSite site) {
         //throw new UnsupportedOperationException("Not yet implemented");
         //first order of business is to remove all hosts from the site
@@ -47,34 +51,38 @@ public class PsObjectShredder {
         // ok, now we can remove the site
         session.delete(site);
     }
-    
+
     public static void delete(Session session, PsMatrix matrix) {
         //throw new UnsupportedOperationException("Not yet implemented");
-        
+
         // first order of business is to remove all hosts from this matrix
         List<PsHost> allHostsInThisMatrix = matrix.getAllHosts();
         Iterator iter = allHostsInThisMatrix.iterator();
-        while(iter.hasNext()){
-            PsHost hostToBeRemoved = (PsHost)iter.next();
-            PsMatrixManipulator.removeHostFromMatrix(session,matrix, hostToBeRemoved);
+        while (iter.hasNext()) {
+            PsHost hostToBeRemoved = (PsHost) iter.next();
+            PsMatrixManipulator.removeHostFromMatrix(session, matrix, hostToBeRemoved);
         }
         // second order of business is to remove the matrix itself
         session.delete(matrix);
     }
-    
+
     public static void delete(Session session, PsCloud cloud) {
         cloud.removeAllSites();
         cloud.removeAllMatrices();
         session.delete(cloud);
     }
-    
-    public static void delete(Session session, PsJob job){
-        session.delete(job);
+
+    public static void delete(Session session, PsJob job) {
+        if (job != null) {
+            session.delete(job);
+        }
     }
-    
-    public static void deletePsJob(Session session, int jobId){
+
+    public static void deletePsJob(Session session, int jobId) {
         //TODO replace this with HQL DELETE command
-        PsJob job = (PsJob)PsDataStore.getJob(session, jobId);
-        delete(session, job);
+        PsJob job = (PsJob) PsDataStore.getJob(session, jobId);
+        if (job != null) {
+            delete(session, job);
+        }
     }
 }
